Add additional querystring parameters to the eventsource URL, via the middleware's script injection.
This allows me to know:
a) when the process started,
b) when the page was loaded ... well, when the middleware finished and is ready to deliver
c) when the JS eventsource was created.

The page load one is the most important one to track, because between the page being rendered and the event source starting to listen/emit events, files could've been changed (yay race condition), but I'd miss them if I just took the time the SSE connection was opened. By using the page load time as the first value of the last_scan, I can treat it as if another client saw a change in the meantime, and replay it so that it actually happens.