Change the patches for templates, patching fewer methods in favour of just Engine.find_template and introducing one for Template._render
find_template should be collecting _everything_ that gets used, and Template._render will try and persist the values it sees _during a request_ so that a page only reloads if one of the template changes broadcast is for a template which has been seen during the request (injected in via middleware). For template changes which cannot be traced to the given page, a prompt is displayed which the user may back out of, which will preclude changes to that template from reloading, for the time being (stored locally in the JS only currently, rather than localStorage/sessionStorage).

Mostly implemented this because I noticed my admin tab was getting refreshed while editing templates in another tab, which is less than ideal.