If I stash a copy of the ACTUAL reloader in use, and smash the tracked files into the .extra_files set, can I avoid polling separately within the SSE view, and perhaps ameliorate the battery usage by not having to scan and sleep a bunch?
I can't see a better way of getting a process/thread local copy of the reloader currently being used, but it's only runserver so this is probably fine.

From here I can maybe somehow(?) emit the changed value to the looper when file_changed fires for one of the extra_files, which is sort of how listen_for_python_changes works, but with a level of indirection and the possibility of an endless loop...? hmmm.