... I don't know how safe annnny of this is. The idea is to have one (daemon?) thread acting like a stripped down statreloader, and every time it sees a change/deletion, pop that into a queue for each listener. Each listener is a SSE request, and ideally the queues are unbounded and delete when the client/request hangs up.
How safe is it to delete queues like that? How safe is it to iterate queues like that? How safe is it to broadcast like that? I dunno.

What I do know is that currently it's problematic because it's increasing (doubling?) the number of seen items per watch iteration (every 0.35s or 1.0s or whatever)