Prompting about 'possibly unrelated' files should now be persisted per URI, so that accepting a prompt to change 'xyz.html' will always refresh it for that page and not prompt again in future.
Similarly, rejecting a prompt for that file will ignore it for that URL in future (as before).

Uses session storage to keep them around between reloads, and unlike the other save/restore abilities, does *not* delete the key after restoring, so that the choices are kept around more permanently and are less intrusive.

Fixes GH #2