Changed the middleware to not set request.urlconf, so that djdt's DebugToolbarMiddleware works OK, instead of throwing a wobbler about needing the 'djdt' namespace.
Instead I just manually screw around with testing the request path and dispatching immediately to the views.

Also attempts to restore the djdt handle on partial reloads (udomdiff only currently) though that seems slightly not-OK, because the CSS is also refreshed, and by the time the init() tries to read the width of the "handle" element, the CSS is not yet applied and so it's full width and then the "top" CSS declaration ends up set to like -900px or whatever. So I just twist it back into shape and hope for the best ...